+++ app/ide-desktop/lib/dashboard/src/components/dashboard/AssetIcon.tsx
@@ -15,8 +15,8 @@ import * as fileIcon from '#/utilities/fileIcon'
 
 /** Props for an {@link AssetIcon}. */
 export interface AssetIconProps {
-  asset: backend.AnyAsset
-  className?: string
+  readonly asset: backend.AnyAsset
+  readonly className?: string
 }
 
 /** Displays a few details of an asset. */