+++ app/ide-desktop/lib/dashboard/src/layouts/dashboard/UpsertSecretModal.tsx
@@ -15,9 +15,9 @@ import type * as backend from '#/services/Backend'
 
 /** Props for a {@link UpsertSecretModal}. */
 export interface UpsertSecretModalProps {
-  id: backend.SecretId | null
-  name: string | null
-  doCreate: (name: string, value: string) => void
+  readonly id: backend.SecretId | null
+  readonly name: string | null
+  readonly doCreate: (name: string, value: string) => void
 }
 
 /** A modal for creating and editing a secret. */